requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.4

numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
xgboost==2.0.3
joblib==1.4.0

shap==0.45.1
numba==0.59.1
llvmlite==0.42.0

psycopg2-binary==2.9.9
jinja2
